fixed #146

+++ src/JsCache/JsCache.h
@@ -18,12 +18,12 @@ namespace iqrf {
     JsCache();
     virtual ~JsCache();
 
-    const StdDriver* getDriver(int id, int ver) const override;
+    const StdDriver* getDriver(int id, double ver) const override;
     const IJsCacheService::Manufacturer* getManufacturer(uint16_t hwpid) const override;
     const IJsCacheService::Product* getProduct(uint16_t hwpid) const override;
     const IJsCacheService::Package* getPackage(uint16_t hwpid, uint16_t hwpidVer, const std::string& os, const std::string& dpa) const override;
     virtual const Package* getPackage(uint16_t hwpid, uint16_t hwpidVer, uint16_t os, uint16_t dpa) const override;
-    std::map<int, std::map<int, std::vector<std::pair<int, int>>>> getDrivers(const std::string& os, const std::string& dpa) const override;
+    std::map<int, std::map<double, std::vector<std::pair<int, int>>>> getDrivers(const std::string& os, const std::string& dpa) const override;
     std::map<int, std::map<int, std::string>> getCustomDrivers(const std::string& os, const std::string& dpa) const override;
     const OsDpa* getOsDpa(int id) const override;
     const OsDpa* getOsDpa(const std::string& os, const std::string& dpa) const override;
